renamed rule section url
