+++ src/views/partials/doc-content.ejs
@@ -150,7 +150,7 @@
 
   <h1 id="rules-section">Rules</h1>
   <% include ../partials/doc-resource-rules %>
-  <% include ../partials/doc-resource-rules-sections %>
+  <% include ../partials/doc-resource-rule-sections %>
 
   <h1 id="common-models-section">Common Models</h1>
   <% include ../partials/doc-resource-common-models %>
