+++ src/views/partials/doc-left-col.ejs
@@ -79,7 +79,7 @@
           <a href="#rules-section">Rules</a>
           <ul>
             <li><a href="#rules">Rules</a></li>
-            <li><a href="#rules-sections">Rule Sections</a></li>
+            <li><a href="#rule-sections">Rule Sections</a></li>
           </ul>
         </li>
