+++ src/views/partials/doc-resource-rule-sections.ejs
@@ -1,15 +1,15 @@
-<h2 id="rules-sections">Rule Sections</h2>
+<h2 id="rule-sections">Rule Sections</h2>
 <p>
     Rule sections represent a sub-heading and text that can be found underneath a rule heading in the SRD.
 </p>
 
-<h3>GET api/rules-sections/{index}</h3>
+<h3>GET api/rule-sections/{index}</h3>
 
 <pre><code>{
     "name": "Saving Throws",
     "index": "saving-throws",
     "desc": "## Saving Throws\n\nA saving throw-also called a save-represents an attempt to resist a spell, a trap, a poison, a disease, or a similar threat. You don't normally decide to make a saving throw; you are forced to make one because your character or monster is at risk of harm.\n\nTo make a saving throw, roll a d20 and add the appropriate ability modifier. For example, you use your Dexterity modifier for a Dexterity saving throw.\n\nA saving throw can be modified by a situational bonus or penalty and can be affected by advantage and disadvantage, as determined by the GM.\n\nEach class gives proficiency in at least two saving throws. The wizard, for example, is proficient in Intelligence saves. As with skill proficiencies, proficiency in a saving throw lets a character add his or her proficiency bonus to saving throws made using a particular ability score. Some monsters have saving throw proficiencies as well.\n\nThe Difficulty Class for a saving throw is determined by the effect that causes it. For example, the DC for a saving throw allowed by a spell is determined by the caster's spellcasting ability and proficiency bonus.\n\nThe result of a successful or failed saving throw is also detailed in the effect that allows the save. Usually, a successful save means that a creature suffers no harm, or reduced harm, from an effect.\n",
-    "url": "/api/rules-sections/saving-throws"
+    "url": "/api/rule-sections/saving-throws"
   }</code></pre>
 
 <h4>Rule Sections</h4>
