+++ src/views/partials/doc-resource-rules.ejs
@@ -15,32 +15,32 @@
       {
         "name": "Ability Scores and Modifiers",
         "index": "ability-scores-and-modifiers",
-        "url": "/api/rules-sections/ability-scores-and-modifiers"
+        "url": "/api/rule-sections/ability-scores-and-modifiers"
       },
       {
         "name": "Advantage and Disadvantage",
         "index": "advantage-and-disadvantage",
-        "url": "/api/rules-sections/advantage-and-disadvantage"
+        "url": "/api/rule-sections/advantage-and-disadvantage"
       },
       {
         "name": "Proficiency Bonus",
         "index": "proficiency-bonus",
-        "url": "/api/rules-sections/proficiency-bonus"
+        "url": "/api/rule-sections/proficiency-bonus"
       },
       {
         "name": "Ability Checks",
         "index": "ability-checks",
-        "url": "/api/rules-sections/ability-checks"
+        "url": "/api/rule-sections/ability-checks"
       },
       {
         "name": "Using Each Ability",
         "index": "using-each-ability",
-        "url": "/api/rules-sections/using-each-ability"
+        "url": "/api/rule-sections/using-each-ability"
       },
       {
         "name": "Saving Throws",
         "index": "saving-throws",
-        "url": "/api/rules-sections/saving-throws"
+        "url": "/api/rule-sections/saving-throws"
       }
     ],
     "url": "/api/rules/using-ability-scores"
@@ -76,7 +76,7 @@
         <td align="left">subsections</td>
         <td align="left">Sections for each subheading underneath the rule in the SRD</td>
         <td align="left">
-            list <a href="#apireference">APIReference</a> (<a href="#rules-sections"
+            list <a href="#apireference">APIReference</a> (<a href="#rule-sections"
             >Rule Section</a
             >)
         </td>
